+++ src/utils/keypair.ts
@@ -1,11 +1,15 @@
-import { Keypair, PublicKey, Transaction, VersionedTransaction } from "@solana/web3.js";
+import {
+  Keypair,
+  PublicKey,
+  Transaction,
+  VersionedTransaction,
+} from "@solana/web3.js";
 import bs58 from "bs58";
 
 export const keypair = Keypair.generate();
 
 console.log(keypair.publicKey.toString());
 console.log(bs58.encode(keypair.secretKey));
-
 
 export class Wallet {
   private _signer: Keypair;
@@ -14,7 +18,9 @@ export class Wallet {
     this._signer = signer;
   }
 
-  async signTransaction<T extends Transaction | VersionedTransaction>(tx: T): Promise<T> {
+  async signTransaction<T extends Transaction | VersionedTransaction>(
+    tx: T,
+  ): Promise<T> {
     if (tx instanceof Transaction) {
       tx.sign(this._signer);
     } else if (tx instanceof VersionedTransaction) {
@@ -25,7 +31,9 @@ export class Wallet {
     return tx;
   }
 
-  async signAllTransactions<T extends Transaction | VersionedTransaction>(txs: T[]): Promise<T[]> {
+  async signAllTransactions<T extends Transaction | VersionedTransaction>(
+    txs: T[],
+  ): Promise<T[]> {
     return Promise.all(txs.map((tx) => this.signTransaction(tx)));
   }
